EDA/PDF_scraper/requirements.txt
tabula-py>=2.0.5
pandas>=1.3

# Notes:
# - tabula-py requires Java (tabula-java). On Windows, install a JDK and ensure `java` is on PATH.
# - If you prefer a pure-Python extractor (no Java), consider `pdfplumber` or `camelot` (camelot also needs ghostscript).
